Handle quest buttons in Gear controller
One of the button values is big and requires the key in ControllerEvent to be a long.
Now you can select stuff in the 3d cursor app.

+++ SXR/SDK/backend_oculus/src/main/java/com/samsungxr/OvrControllerReader.java
@@ -50,7 +50,7 @@ final class OvrControllerReader extends SXRGearCursorController.ControllerReader
         event.position.set(readbackBuffer[controllerID].get(INDEX_POSITION),
                 readbackBuffer[controllerID].get(INDEX_POSITION + 1),
                 readbackBuffer[controllerID].get(INDEX_POSITION + 2));
-        event.key = (int) readbackBuffer[controllerID].get(INDEX_BUTTON);
+        event.key = (long) readbackBuffer[controllerID].get(INDEX_BUTTON);
 
         controllerEvents.add(event);
     }
